python, eccodes and magics versions updated (#40)
* python, eccodes and magics versions updated
- minimum supported python version is now 3.10
- support for python up to version 3.14 added
- eccodes bumped to latest available 2.46.3
- magics version bumped to latest available 4.16.1
- hdf5 lib bumped to 1.14.6
- ecmwflibs bumped to 0.7.0

* upload artifact action updated to v4

* updated build commands to fix "GLIBC_2.27/2.28 not found" error

* fixing windows hdf5 error

* temporary venv added for macos builds

* macox build fix

* build step for libaec added

* fixing various build errors

* fix for linux build (libaec cmake package not found)

* another linux build fix

* added mozilla header to avoid timeouts downloading lgpl license

* removed hdf-5 system lib, changed libjpeg to build from source, licences updated

* fixed hdf 5 version string

* lgpl fallback to file added, license fetching error handling improved

* disabled netcdf tests for linux build

* another attempt at fixing build errors in linux

* libjasper added to linux source build

* fixes for windows build and python 3.12+

* fix for windows x64 python 3.14 tests

+++ scripts/wheel-linux.sh
@@ -11,12 +11,23 @@ set -eaux
 
 version=$(echo $1| sed 's/\.//')
 
+pybin=$(ls -1d /opt/python/cp${version}-cp${version}*/bin/python3 2>/dev/null | head -1)
+if [[ -z "$pybin" ]]
+then
+	pybin=$(ls -1d /opt/python/cp${version}t-cp${version}t*/bin/python3 2>/dev/null | head -1)
+fi
+if [[ -z "$pybin" ]]
+then
+	echo "Cannot find Python binary for cp${version} under /opt/python"
+	exit 1
+fi
+
 TOPDIR=$(/bin/pwd)
 
-LD_LIBRARY_PATH=$TOPDIR/install/lib:$TOPDIR/install/lib64:$LD_LIBRARY_PATH
+export LD_LIBRARY_PATH=$TOPDIR/install/lib:$TOPDIR/install/lib64:${LD_LIBRARY_PATH:-}
 
 rm -fr dist wheelhouse
-/opt/python/cp${version}-cp${version}*/bin/python3 setup.py bdist_wheel
+$pybin setup.py bdist_wheel
 
 # Do it twice to get the list of libraries
 
@@ -27,6 +38,6 @@ pip3 install -r tools/requirements.txt
 python3 ./tools/copy-licences.py libs
 
 rm -fr dist wheelhouse
-/opt/python/cp${version}-cp${version}*/bin/python3 setup.py bdist_wheel
+$pybin setup.py bdist_wheel
 auditwheel repair dist/*.whl
 rm -fr dist
